agency-purchases-ms/requirements.txt
django==3.2.7
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-heroku==0.3.1
dj-database-url==0.5.0
python-dotenv==0.19.0
psycopg2==2.9.1
gunicorn==20.1.0
whitenoise==5.3.0